+++ src/scope_alpaca.cpp
@@ -45,7 +45,7 @@ ScopeAlpaca::ScopeAlpaca()
 {
     m_client = nullptr;
     m_host = pConfig->Profile.GetString("/alpaca/host", _T("localhost"));
-    m_port = pConfig->Profile.GetLong("/alpaca/port", 6800);
+    m_port = pConfig->Profile.GetLong("/alpaca/port", 0);
     m_deviceNumber = pConfig->Profile.GetLong("/alpaca/telescope_device", 0);
     m_Name = wxString::Format("Alpaca Mount [%s:%ld/%ld]", m_host, m_port, m_deviceNumber);
 
@@ -84,12 +84,12 @@ bool ScopeAlpaca::Connect()
         }
 
         // If not configured open the setup dialog
-        if (m_host == _T("localhost") && m_port == 6800 && m_deviceNumber == 0)
+        if (m_port == 0)
         {
             SetupDialog();
             // Reload values after dialog
             m_host = pConfig->Profile.GetString("/alpaca/host", _T("localhost"));
-            m_port = pConfig->Profile.GetLong("/alpaca/port", 6800);
+            m_port = pConfig->Profile.GetLong("/alpaca/port", 0);
             m_deviceNumber = pConfig->Profile.GetLong("/alpaca/telescope_device", 0);
         }